connectredismongodb/requirements.txt
django
djangorestframework
djongo
pymongo==3.12.1
django-redis
dnspython